restructure: HOC 폴더 생성 및 타입 위치 변경
