+++ src/components/shared/input/type.ts
@@ -1,4 +1,4 @@
-type InputType =
+export type InputType =
   | 'text'
   | 'email'
   | 'password'
@@ -7,17 +7,3 @@ type InputType =
   | 'tel'
   | 'url'
   | 'date';
-
-export interface InputProps
-  extends React.InputHTMLAttributes<HTMLInputElement> {
-  type?: InputType;
-  id: string;
-  ariaLabel?: string;
-}
-
-export interface WithLabelProps
-  extends React.LabelHTMLAttributes<HTMLLabelElement> {
-  id: string;
-  name: string;
-  labelClassName?: string;
-}
